refactor: Merge MoECrossEntropyLoss into modalities core
